# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Jikū API",
    "description" : "Event invitation, ticketing, RSVP and check-in API.",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "security" : [
    {
      "bearer-jwt" : [ ]
    }
  ],
  "paths" : {
...
    "/api/v1/events/{eventId}/payments/manual" : {
      "get" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "current",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "request",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/billing/invoices/{invoiceId}/credit-note" : {
      "post" : {
        "tags" : [
          "invoice-controller"
        ],
        "operationId" : "creditNote",
        "parameters" : [
          {
            "name" : "invoiceId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/invoices/payments/{paymentId}" : {
      "post" : {
        "tags" : [
          "invoice-controller"
        ],
        "operationId" : "issueForPayment",
        "parameters" : [
          {
            "name" : "paymentId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/billing/invoices/{invoiceId}" : {
      "get" : {
        "tags" : [
          "invoice-controller"
        ],
        "operationId" : "get_5",
        "parameters" : [
          {
            "name" : "invoiceId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InvoiceDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/invoices/{invoiceId}/document" : {
      "get" : {
        "tags" : [
          "invoice-controller"
        ],
        "operationId" : "document",
        "parameters" : [
          {
            "name" : "invoiceId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ManualPaymentRequest" : {
        "type" : "object",
        "properties" : {
          "tier" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "tier"
        ]
      },
      "ManualPaymentInstructions" : {
        "type" : "object",
        "properties" : {
          "paymentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reference" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "payee" : {
            "$ref" : "#/components/schemas/PayeeDetails"
          }
        }
      },
      "PayeeDetails" : {
        "type" : "object",
        "properties" : {
          "payeeName" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyNumber" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyOperator" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "bankDetails" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
...
      "InvoiceDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "invoiceNumber" : {
            "type" : "string"
          },
          "documentType" : {
            "type" : "string"
          },
          "correctedInvoiceId" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "uuid"
          },
          "buyer" : {
            "$ref" : "#/components/schemas/InvoiceParty"
          },
          "seller" : {
            "$ref" : "#/components/schemas/InvoiceParty"
          },
          "currency" : {
            "type" : "string"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InvoiceLineView"
            }
          },
          "subtotalMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "taxLabel" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "taxRate" : {
            "type" : "number"
          },
          "taxAmountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "issueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "issuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "paymentId" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "uuid"
          }
        }
      },
      "InvoiceLineView" : {
        "type" : "object",
        "properties" : {
          "lineNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "description" : {
            "type" : "string"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unitPriceMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineTotalMinor" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "InvoiceParty" : {
        "type" : "object",
        "properties" : {
          "legalName" : {
            "type" : "string"
          },
          "addressLine" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "city" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "country" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "registrationNumber" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "taxIdentifier" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}